% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_distribution.R
\docType{data}
\name{age_distribution}
\alias{age_distribution}
\title{Age distribution of RSV hospitalizations}
\format{
\subsection{\code{age_distribution}}{

A data frame with 65 rows and 3 columns:
\describe{
\item{state}{state abbreviation}
\item{age_group}{age group (<6 months, 6-11 months, 1-4 years, 5-59 years, 60+ years)}
\item{proportion}{proportion of all hospitalizations from the 2018-2019 season to the 2023-2024 season}
...
}
}
}
\source{
\url{https://github.com/midas-network/rsv-scenario-modeling-hub/tree/main/target-data}
}
\usage{
age_distribution
}
\description{
Age distribution of RSV hospitalizations for 12 states participating in RSVNet
}
\keyword{datasets}
