% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_scenarios.R
\name{display_scenarios}
\alias{display_scenarios}
\title{Launch Shiny App to display scenario projections (function)}
\usage{
display_scenarios()
}
\description{
This function launches the Shiny App included in the package.
The Shiny App allows users to upload the saved results from the package functions and display scenario projections.
In order for the Shiny App to work, make sure that one of the scenarios is called "Counterfactual" and that projection_intervals = TRUE
}
