% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Download and format data for RSV dynamic transmission model}
\usage{
get_data(state_or_county, state_abbr, county_name)
}
\arguments{
\item{state_or_county}{describe if you are modeling for a "state" or a "county"}

\item{state_abbr}{The two letter state abbreviation (you need this even if you are modeling a county)}

\item{county_name}{The name of the county (if modeling a state list NULL here)}
}
\value{
a list of values
}
\description{
Download and format data for RSV dynamic transmission model
}
\examples{
data_california = get_data(state_or_county="state",state_abbr="CA",county_name=NULL)
data_kingcounty = get_data(state_or_county="county",state_abbr="WA",county_name="King")
}
