% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSIRS_immunization_dynamics.R
\name{MSIRS_immunization_dynamics}
\alias{MSIRS_immunization_dynamics}
\title{MSIRS dynamic transmission model for RSV}
\usage{
MSIRS_immunization_dynamics(t, y, parms)
}
\arguments{
\item{t}{A vector of numbers representing the number of weeks/months to run the model}

\item{y}{A vector of starting values for model compartments}

\item{parms}{A list of fixed parameter values}
}
\value{
A matrix of values.
}
\description{
MSIRS dynamic transmission model for RSV
}
\examples{
dat = get_data(state_or_county="state",state_abbr="CA",county_name=NULL)
parmset=dat[[1]]
yinit=dat[[2]]
yinit.vector=dat[[3]]
results = MSIRS_immunization_dynamics(t=seq(1,100,by=1),y=yinit.vector,parms=parmset)
}
