% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_distribution.R
\docType{data}
\name{age_distribution}
\alias{age_distribution}
\title{Age distribution of RSV hospitalizations}
\format{
\subsection{\code{age_distribution}}{

A data frame with 55 rows and 3 columns:
\describe{
\item{state}{state name}
\item{age_group}{age group (<6 months, 6-11 months, 1-4 years, 5-64 years, 65+ years)}
\item{proportion}{proportion of all hospitalizations during the 2018-2019 and 2029-2020 seasons}
...
}
}
}
\source{
\url{https://data.cdc.gov/Public-Health-Surveillance/Rates-of-Laboratory-Confirmed-RSV-COVID-19-and-Flu/kvib-3txy/about_data}
}
\usage{
age_distribution
}
\description{
Age distribution of RSV hospitalizations for 12 states participating in RSVNet
}
\keyword{datasets}
