% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSIRS_immunization_dynamics.R
\name{MSIRS_immunization_dynamics}
\alias{MSIRS_immunization_dynamics}
\title{MSIRS dynamic transmission model for RSV}
\usage{
MSIRS_immunization_dynamics(times, y, parms)
}
\arguments{
\item{times}{A numeric vector of times}

\item{y}{A vector of starting values for model compartments}

\item{parms}{A list of fixed parameter values}
}
\value{
A matrix of values.
}
\description{
MSIRS dynamic transmission model for RSV
}
\examples{

dat = get_data(state_or_county="state",state_abbr="CA",county_name=NULL)
parmset=dat[[1]]
yinit.vector=dat[[3]]
fit_times = seq(1,100,by=1)
parms=c(parmset,baseline.txn.rate=7,b1=.12,phi=3.2)
results <- MSIRS_immunization_dynamics(times=fit_times, y=yinit.vector,parms=parms)
}
