% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\docType{data}
\name{timeseries}
\alias{timeseries}
\title{Time series of RSV hospitalizations (data frame)}
\format{
\subsection{\code{timeseries}}{

A data frame with 1274 rows and 3 columns:
\describe{
\item{date}{MMWR week}
\item{state}{state name (California, Georgia, Maryland, Minnesota, New York, Tennessee, Oregon)}
\item{value}{estimated number of RSV hospitalizations}
...
}
}
}
\source{
\url{https://data.cdc.gov/Public-Health-Surveillance/Weekly-Rates-of-Laboratory-Confirmed-RSV-Hospitali/29hc-w46k/about_data}
}
\usage{
timeseries
}
\description{
Weekly RSV hospitalizations (all ages).
Data is available for the 7 states that began participation in RSV-Net in 2016.
RSV-Net provides data on RSV hospitalization rates.
We have converted these to an approximate number of hospitalizations for use in the model.
RSV-Net did not include data for children until 2018.
Simulated data for children has been added based on the weekly average RSV hospitalization rate among children during the 2018-19 and 2019-20 RSV seasons.
Data have been inflated to adjust for increases in RSV testing and reporting after 2020.
}
\keyword{datasets}
